Moved and updated search_target_token_paths in refs/base.py
